--- source/common/crypto/BUILD
@@ -1,23 +1,0 @@
-licenses(["notice"])  # Apache 2
-
-load(
-    "//bazel:envoy_build_system.bzl",
-    "envoy_cc_library",
-    "envoy_package",
-)
-
-envoy_package()
-
-envoy_cc_library(
-    name = "utility_lib",
-    srcs = ["utility.cc"],
-    hdrs = ["utility.h"],
-    external_deps = [
-        "ssl",
-    ],
-    deps = [
-        "//include/envoy/buffer:buffer_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:stack_array",
-    ],
-)
