+++ source/extensions/filters/http/common/aws/BUILD
@@ -28,9 +28,9 @@ envoy_cc_library(
         "//source/common/common:hex_lib",
         "//source/common/common:logger_lib",
         "//source/common/common:utility_lib",
-        "//source/common/crypto:utility_lib",
         "//source/common/http:headers_lib",
         "//source/common/singleton:const_singleton",
+        "//source/extensions/transport_sockets/tls:utility_lib",
     ],
 )
