+++ source/extensions/filters/http/lua/wrappers.h
@@ -3,9 +3,8 @@
 #include "envoy/http/header_map.h"
 #include "envoy/stream_info/stream_info.h"
 
-#include "common/crypto/utility.h"
-
 #include "extensions/filters/common/lua/lua.h"
+#include "extensions/transport_sockets/tls/utility.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -208,7 +207,8 @@ private:
  */
 class PublicKeyWrapper : public Filters::Common::Lua::BaseLuaObject<PublicKeyWrapper> {
 public:
-  PublicKeyWrapper(Envoy::Common::Crypto::PublicKeyPtr key) : public_key_(std::move(key)) {}
+  PublicKeyWrapper(Envoy::Extensions::TransportSockets::Tls::PublicKeyPtr key)
+      : public_key_(std::move(key)) {}
   static ExportedFunctions exportedFunctions() { return {{"get", static_luaGet}}; }
 
 private:
@@ -218,7 +218,7 @@ private:
    */
   DECLARE_LUA_FUNCTION(PublicKeyWrapper, luaGet);
 
-  Envoy::Common::Crypto::PublicKeyPtr public_key_;
+  Envoy::Extensions::TransportSockets::Tls::PublicKeyPtr public_key_;
 };
 
 } // namespace Lua
