+++ source/extensions/transport_sockets/tls/BUILD
@@ -110,7 +110,9 @@ envoy_cc_library(
         "ssl",
     ],
     deps = [
+        "//include/envoy/buffer:buffer_interface",
         "//source/common/common:assert_lib",
+        "//source/common/common:stack_array",
         "//source/common/common:utility_lib",
     ],
 )
