+++ source/extensions/transport_sockets/tls/utility.h
@@ -1,16 +1,37 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
+#include "envoy/buffer/buffer.h"
+
 #include "common/common/utility.h"
 
+#include "absl/strings/string_view.h"
+#include "openssl/evp.h"
 #include "openssl/ssl.h"
 
 namespace Envoy {
 namespace Extensions {
 namespace TransportSockets {
 namespace Tls {
+
+struct VerificationOutput {
+  /**
+   * Verification result. If result_ is true, error_message_ is empty.
+   */
+  bool result_;
+
+  /**
+   * Error message when verification failed.
+   * TODO(crazyxy): switch to absl::StatusOr when available
+   */
+  std::string error_message_;
+};
+
+typedef bssl::UniquePtr<EVP_PKEY> PublicKeyPtr;
+
 namespace Utility {
 
 /**
@@ -57,6 +78,43 @@ SystemTime getValidFrom(const X509& cert);
  * @return time after which the certificate expires.
  */
 SystemTime getExpirationTime(const X509& cert);
+
+/**
+ * Computes the SHA-256 digest of a buffer.
+ * @param buffer the buffer.
+ * @return a vector of bytes for the computed digest.
+ */
+std::vector<uint8_t> getSha256Digest(const Buffer::Instance& buffer);
+
+/**
+ * Computes the SHA-256 HMAC for a given key and message.
+ * @param key the HMAC function key.
+ * @param message message data for the HMAC function.
+ * @return a vector of bytes for the computed HMAC.
+ */
+std::vector<uint8_t> getSha256Hmac(const std::vector<uint8_t>& key, absl::string_view message);
+
+/**
+ * Verify cryptographic signatures.
+ * @param hash hash function(including SHA1, SHA224, SHA256, SHA384, SHA512)
+ * @param key pointer to public key
+ * @param signature signature
+ * @param text clear text
+ * @return If the result_ is true, the error_message_ is empty; otherwise,
+ * the error_message_ stores the error message
+ */
+const VerificationOutput verifySignature(absl::string_view hash, EVP_PKEY* key,
+                                         const std::vector<uint8_t>& signature,
+                                         const std::vector<uint8_t>& text);
+
+/**
+ * Import public key.
+ * @param key key string
+ * @return pointer to public key
+ */
+PublicKeyPtr importPublicKey(const std::vector<uint8_t>& key);
+
+const EVP_MD* getHashFunction(absl::string_view name);
 
 } // namespace Utility
 } // namespace Tls
