--- test/common/crypto/BUILD
@@ -1,22 +1,0 @@
-licenses(["notice"])  # Apache 2
-
-load(
-    "//bazel:envoy_build_system.bzl",
-    "envoy_cc_test",
-    "envoy_package",
-)
-
-envoy_package()
-
-envoy_cc_test(
-    name = "utility_test",
-    srcs = [
-        "utility_test.cc",
-    ],
-    external_deps = ["ssl"],
-    deps = [
-        "//source/common/buffer:buffer_lib",
-        "//source/common/common:hex_lib",
-        "//source/common/crypto:utility_lib",
-    ],
-)
